Use big ints in migrations

+++ src/database/migrations/2022_10_29_000002_create_accounting_journals_table.php
@@ -11,12 +11,12 @@ class CreateAccountingJournalsTable extends Migration
     public function up(): void
     {
         Schema::create('accounting_journals', function (Blueprint $table) {
-            $table->increments('id');
-            $table->unsignedInteger('ledger_id')->nullable();
+            $table->bigIncrements('id');
+            $table->unsignedBigInteger('ledger_id')->nullable();
             $table->bigInteger('balance');
             $table->string('currency', 5);
             $table->string('morphed_type', 32);
-            $table->integer('morphed_id');
+            $table->bigInteger('morphed_id');
             $table->timestamps();
         });
     }
